refactor(next): move RouterAdapter to adapters/ directory
Groups framework adapter implementations under packages/next/src/adapters/
instead of mixing them into elements/ (UI primitives) and utilities/ (misc
helpers). The Next adapter is neither — it bridges next/navigation into the
RouterAdapterContext contract from payload core. Sets up the structure for
the upcoming ServerAdapter to live alongside as adapters/server.ts.
