+++ packages/next/src/layouts/Root/index.tsx
@@ -10,8 +10,8 @@ import { cookies as nextCookies } from 'next/headers.js'
 import { applyLocaleFiltering } from 'payload/shared'
 import React, { Suspense } from 'react'
 
+import { NextRouterAdapter } from '../../adapters/router.js'
 import { getNavPrefs } from '../../elements/Nav/getNavPrefs.js'
-import { NextRouterAdapter } from '../../elements/RouterAdapter/index.js'
 import { getRequestHighContrast } from '../../utilities/getRequestHighContrast.js'
 import { getRequestTheme } from '../../utilities/getRequestTheme.js'
 import { initReq } from '../../utilities/initReq.js'
